+++ curve25519-dalek/CHANGELOG.md
@@ -43,6 +43,7 @@ major series.
   version.
 * Relax the `zeroize` dependency to `^1`
 * Update the edition from 2015 to 2021
+* Add implementations of the `ff` and `group` traits, behind the `group` feature flag.
 
 ## 3.x series